server/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.0
alembic>=1.13.0

# Embeddings and ML
sentence-transformers>=2.3.1
torch>=2.1.0
transformers>=4.38.0
numpy>=1.26.0

# LLM Providers
openai>=1.10.0
anthropic>=0.8.0
google-genai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
aiosqlite>=0.19.0

# Logging and monitoring
structlog>=24.1.0
